Fix cnt overflow issue

+++ ModelOptimizations/DlQuantization/include/DlQuantization/Fp16Quantization.hpp
@@ -40,6 +40,7 @@
 #define FP16_QUANTIZATION_HPP
 
 #include "DlQuantization/Quantization.hpp"
+#include <cstdint>
 
 namespace DlQuantization
 {
@@ -49,7 +50,7 @@ namespace DlQuantization
  * @param cnt total size of input tensor
  * @param out pointer to the output tensor
  */
-void quantizeDequantizeFp16Gpu(const float* in, int cnt, float* out, void* stream = nullptr);
+void quantizeDequantizeFp16Gpu(const float* in, uint64_t cnt, float* out, void* stream = nullptr);
 
 }   // namespace DlQuantization
